+++ src/include/backup/walsummary.h
@@ -42,8 +42,8 @@ extern File OpenWalSummaryFile(WalSummaryFile *ws, bool missing_ok);
 extern void RemoveWalSummaryIfOlderThan(WalSummaryFile *ws,
 										time_t cutoff_time);
 
-extern int	ReadWalSummary(void *wal_summary_io, void *data, int length);
-extern int	WriteWalSummary(void *wal_summary_io, void *data, int length);
+extern size_t ReadWalSummary(void *wal_summary_io, void *data, size_t length);
+extern size_t WriteWalSummary(void *wal_summary_io, void *data, size_t length);
 extern void ReportWalSummaryError(void *callback_arg, char *fmt, ...) pg_attribute_printf(2, 3);
 
 #endif							/* WALSUMMARY_H */
